Register Pod process in a regine pool
Before this commit regine registry mananges only ets table of
pods and associated data with these pods. This leads to situation
when pod is deleted/restarted - data was not removed fully from the
ets registry table.

This commits adds registration of a new pod process in the regine
pool. So when the process will be terminated - all the data related
to it should be cleaned-up.